Update functions.md
The execution of netcat to interact with the client sockaddr_in app, it's not correct.
Instead of "nc -l 5000", it should be "nc -lp 5000" 
It means the the parameter p is missing to define that it the port we want to define. Without that parameter, it listens in a random port.