Improve Android documentation
The Android documentation could use an enhancement for instructions on decompressing XZ files. It took me a few minutes to find that it needed to be decompressed as well as the correct utility. Other users have run into this in the past too.
https://github.com/frida/frida/issues/364
https://github.com/frida/frida/issues/367

It's a bit more verbose but can help beginners like myself.